lib/modules: add mkRenamedFreeFormOptionModule
Return a module that causes a warning to be shown if the specified "from"
option is being read, while still returning the value of the "to" option.
This can be used to rename free-form options, while providing backward
compatibility for consumers and blocking writes to the option.
It's i.e. useful for renaming `system.build` options, for example:

  mkRenamedFreeFormOptionModule [ "system" "build" "linodeImage" ] [ "system" "build" "images" ]

warns if system.build.linodeImage gets read, but still evaluates as expected.